+++ contributions/api/controllers/configs.py
@@ -28,11 +28,11 @@ CONTRIBUTION_URL_PREFIX = os.getenv('CONTRIBUTION_URL_PREFIX', '')
 DEBUG = bool(os.getenv('DEBUG', 'False') == 'True')
 CORS_ENABLED = bool(os.getenv('CORS_ENABLED', 'False') == 'True')
 
-
 CONTRIBUTION_DB_NAME = os.getenv('CONTRIBUTION_DB_NAME', 'contributions_db')
 CONTRIBUTION_COLL_NAME = os.getenv('CONTRIBUTION_COLL_NAME', 'contributions')
-CAPABILITY_COLL_NAME = os.getenv('CAPABILITY_COLL_NAME', 'capabilities')
+REVIEWER_COLL_NAME = os.getenv('REVIEWER_COLL_NAME', 'reviewers')
 
 FIELD_OBJECTID = '_id'
 FIELD_NAME = 'name'
 
+ADMIN_USERS = os.getenv('ADMIN_USERS', '')