+++ contributions/catalog/catalog_rest_service.py
@@ -4,7 +4,7 @@ from time import gmtime
 
 from flask import Flask, redirect, url_for, render_template, request, session
 from requests_oauthlib import OAuth2Session
-
+from controllers.auth import login_required
 from controllers.config import Config as cfg
 from controllers.contribute import bp as contribute_bp
 from db import init_app
@@ -37,17 +37,22 @@ init_app(app)
 app.register_blueprint(contribute_bp)
 
 
-@app.route("/")
+@app.route("/", methods=["GET"])
 def index():
+    return render_template('contribute/home.html')
+
+
+@app.route("/login")
+def login():
     """Step 1: Get the user identify for authentication.
     """
     # print("Step 1: User Authorization")
-    github = OAuth2Session(cfg.CLIENT_ID)
+    github = OAuth2Session(cfg.GITHUB_CLIENT_ID)
     authorization_url, state = github.authorization_url(cfg.AUTHORIZATION_BASE_URL)
 
     # State is used to prevent CSRF.
     session['oauth_state'] = state
-    print(session)
+
     return redirect(authorization_url)
 
 
@@ -58,24 +63,25 @@ def callback():
     """ Step 3: Retrieving an access token.
     """
     # print("Step 3: Retrieving an access token")
-    github = OAuth2Session(cfg.CLIENT_ID, state=session['oauth_state'])
-    token = github.fetch_token(cfg.TOKEN_URL, client_secret=cfg.CLIENT_SECRET,
+    github = OAuth2Session(cfg.GITHUB_CLIENT_ID, state=session['oauth_state'])
+    token = github.fetch_token(cfg.TOKEN_URL, client_secret=cfg.GITHUB_CLIENT_SECRET,
                                authorization_response=request.url)
     session['oauth_token'] = token
+
     return redirect(url_for('.profile'))
 
-@app.route("/profile", methods=["GET"])
+@app.route("/contribute/profile", methods=["GET"])
 def profile():
     """Fetching a protected resource using an OAuth 2 token.
     Parsing the username to the seesion dict, to the templates to display.
     """
     # print("Fetching a protected resource using an OAuth 2 token")
-    github = OAuth2Session(cfg.CLIENT_ID, token=session['oauth_token'])
+    github = OAuth2Session(cfg.GITHUB_CLIENT_ID, token=session['oauth_token'])
     resp = github.get('https://api.github.com/user')
-    print(resp.json())
     session["username"] = resp.json()["login"]
     session['name'] = resp.json()["name"]
-    return render_template('contribute/home.html', user=session["name"])
+
+    return render_template('contribute/home.html', user=session["name"], token=session['oauth_token']['access_token'])
 
 
 if __name__ == '__main__':