+++ contributions/catalog/models/capability_utilities.py
@@ -1,5 +1,22 @@
+#  Copyright 2021 Board of Trustees of the University of Illinois.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+import uuid
+
 def init_capability():
-    d = {'name': '',
+    d = {'id': '',
+         'name': '',
          'description': '',
          'icon': None,
          'apiDocUrl': None,
@@ -36,6 +53,8 @@ def to_capability(d):
             capability_list.append(init_capability())
 
     for i, capability in enumerate(capability_list):
+        cap_id = str(uuid.uuid4())
+        capability['id'] = cap_id
         env_k, env_v = d['environmentVariables_key'], d['environmentVariables_value']
         for k, v in list(zip(env_k, env_v)):
             capability["deploymentDetails"]['environmentVariables'].append({'key': k, 'value': v})