+++ contributions/catalog/models/talent_utilities.py
@@ -1,8 +1,23 @@
-from datetime import date
+#  Copyright 2021 Board of Trustees of the University of Illinois.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 
+from datetime import date
+import uuid
 
 def init_talent():
     d = {
+        'id': '',
         "name": '',
         "shortDescription": '',
         "longDescription": '',
@@ -35,9 +50,11 @@ def to_talent(d):
             talent_list.append(init_talent())
 
     for i, talent in enumerate(talent_list):
+        tal_id = str(uuid.uuid4())
+        talent['id'] = tal_id
+
         for k, v in d.items():
             print(k, v)
-
             if "minUserPrivacyLevel" in k:
                 if len(v[i]) > 0:
                     talent_list[i]["minUserPrivacyLevel"] = int(v[i])