+++ release_contributions.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+source release_base_script.sh
+
+###### CONTRIBUTIONS BUILDING BLOCK ######
+docker build --pull -f contributions/Dockerfile -t ${PROJECT_NAME}/contributions-building-block:${VERSION} .
+docker tag ${PROJECT_NAME}/contributions-building-block:${VERSION} 779619664536.dkr.ecr.us-east-2.amazonaws.com/${PROJECT_NAME}/contributions-building-block:${VERSION}
+docker push 779619664536.dkr.ecr.us-east-2.amazonaws.com/${PROJECT_NAME}/contributions-building-block:${VERSION}